+++ source/common/modules/markdown-editor/tooltips/file-preview.ts
@@ -16,7 +16,10 @@
 import { syntaxTree } from '@codemirror/language'
 import { hoverTooltip, type EditorView, type Tooltip } from '@codemirror/view'
 import { trans } from '@common/i18n-renderer'
+import { md2html } from '@common/modules/markdown-utils/markdown-to-html'
 import formatDate from '@common/util/format-date'
+import { CITEPROC_MAIN_DB } from '@dts/common/citeproc'
+import sanitizeHtml from 'sanitize-html'
 
 const ipcRenderer = window.ipc
 
@@ -69,13 +72,24 @@ function getPreviewElement (metadata: [string, string, number, number], linkCont
   const wrapper = document.createElement('div')
   wrapper.classList.add('editor-note-preview')
 
-  const title = document.createElement('h4')
+  const title = document.createElement('p')
   title.classList.add('filename')
   title.textContent = metadata[0]
 
   const content = document.createElement('div')
   content.classList.add('note-content')
-  content.textContent = metadata[1]
+  const html = md2html(metadata[1], window.getCitationCallback(CITEPROC_MAIN_DB))
+  content.innerHTML = sanitizeHtml(html, {
+    // These options basically translate into: Allow nothing but bare metal tags
+    allowedTags: sanitizeHtml.defaults.allowedTags.concat(['img']),
+    disallowedTagsMode: 'escape',
+    allowedIframeDomains: [],
+    allowedIframeHostnames: [],
+    allowedScriptDomains: [],
+    allowedSchemes: [],
+    allowedScriptHostnames: [],
+    allowVulnerableTags: false
+  })
 
   const meta = document.createElement('div')
   meta.classList.add('metadata')
@@ -127,7 +141,9 @@ function getPreviewElement (metadata: [string, string, number, number], linkCont
   }
 
   wrapper.appendChild(title)
+  wrapper.appendChild(document.createElement('hr'))
   wrapper.appendChild(content)
+  wrapper.appendChild(document.createElement('hr'))
   wrapper.appendChild(meta)
   wrapper.appendChild(actions)